refactor: extract credit usage body
